Tests/LibWeb/Text/expected/css/font-face-media-query-change-during-rendering.txt

Initial media query matches: true
Initial font face status: loading
Final media query matches: false
Font face remains connected: false
PASS: Media query changed while a CSS font was loading
